refactor: split background upload extension workflows
Move destination, job, and result handling into focused extension files and add retry processing for background upload jobs.
